+++ utils/widgets/vectorWidget.cpp
@@ -92,7 +92,7 @@ vector<double> DoubleVectorWidget::value()
 
     for (size_t i = 0; i < toReturn.size(); i++)
     {
-        QListWidgetItem *item = ui->listWidget->item(i);
+        QListWidgetItem *item = ui->listWidget->item((int)i);
         QDoubleSpinBox *dw = static_cast<QDoubleSpinBox *>(ui->listWidget->itemWidget(item));
         toReturn[i] = dw->value();
     }
@@ -101,10 +101,10 @@ vector<double> DoubleVectorWidget::value()
 
 void DoubleVectorWidget::setValue(const vector<double> &value)
 {
-    resize(value.size());
+    resize((int)value.size());
     for (size_t i = 0; i < value.size(); i++)
     {
-        QListWidgetItem *item = ui->listWidget->item(i);
+        QListWidgetItem *item = ui->listWidget->item((int)i);
         QDoubleSpinBox *dw = static_cast<QDoubleSpinBox *>(ui->listWidget->itemWidget(item));
         dw->setValue(value[i]);
     }
@@ -115,7 +115,7 @@ void DoubleVectorWidget::setMaximum(double value)
     mMaximum = value;
     for (size_t i = 0; i < (size_t)ui->listWidget->count(); i++)
     {
-        QListWidgetItem *item = ui->listWidget->item(i);
+        QListWidgetItem *item = ui->listWidget->item((int)i);
         QDoubleSpinBox *dw = static_cast<QDoubleSpinBox *>(ui->listWidget->itemWidget(item));
         dw->setMaximum(mMaximum);
     }
@@ -126,7 +126,7 @@ void DoubleVectorWidget::setMinimum(double value)
     mMinimum = value;
     for (size_t i = 0; i < (size_t)ui->listWidget->count(); i++)
     {
-        QListWidgetItem *item = ui->listWidget->item(i);
+        QListWidgetItem *item = ui->listWidget->item((int)i);
         QDoubleSpinBox *dw = static_cast<QDoubleSpinBox *>(ui->listWidget->itemWidget(item));
         dw->setMinimum(mMinimum);
     }
